# Improve binding redirects performance