--- a/src/Paket.Core/Installation/BindingRedirects.fs
+++ b/src/Paket.Core/Installation/BindingRedirects.fs
@@ -7,6 +7,7 @@ open System.Xml.Linq
 open System.IO
 open Paket.Xml.Linq
 open System.Xml.XPath
+open System.Reflection
 
 /// Represents a binding redirection
 type BindingRedirect = 
@@ -196,9 +197,8 @@ let applyBindingRedirectsToFolder isFirstGroup createNewBindingFiles cleanBindin
         applyBindingRedirects p
 
 /// Calculates the short form of the public key token for use with binding redirects, if it exists.
-let getPublicKeyToken (assembly:Mono.Cecil.AssemblyDefinition) =
-    ("", assembly.Name.PublicKeyToken)
-    ||> Array.fold(fun state b -> state + b.ToString("X2"))
-    |> function
-    | "" -> None
-    | token -> Some (token.ToLower())
+let getPublicKeyToken (assembly:AssemblyName) =
+    assembly.GetPublicKeyToken()
+    |> Option.ofObj
+    |> Option.map (Array.fold(fun state b -> state + b.ToString("X2")) "")
+    |> Option.bind (function | "" -> None | token -> Some (token.ToLower()))