This PR changes from `Mono.Cecil` to `AssemblyReader` to read the assemblies to find out if binding redirects is needed.

`paket install` results:

|Project Size|Before PR|After PR|
|-----------|-----------|-------|
|Small Solution (4 projects)|1 second|1 second|
|Medium Solution (29 projects)|26 seconds|4 seconds|
|Large Solution (72 projects)|1 minute, 33 seconds|13 seconds|