@forki 2018-06-07
IIRC we used Cecil because of mono bugs. Maybe they are fixed now and we can try again