@matthid 2018-06-07
Is this the same API as in 
https://github.com/fsprojects/Paket/pull/3223?

Why do we need no new dependency here?

Is this reflection loading the assembly or just reading it?

 /cc @viktor-svub