@viktor-svub 2018-06-07
seems similar to #3223 but it's very different in both intent and implementation:

improve binding redirects performance #3239 | removing Mono.Cecil #3223 
--- | ---
much cleaner / simpler refactoring | bigger refactoring to different API 
targets and really improves performance | targets removal of 3rd-party dependency
does not remove the original dependency | may improve performance as side-effect
hard-coded binary metadata reader | standard ECMA-335 metadata reader
custom / 3rd-party / legacy? `AssemblyReader.fs` | Microsoft `System.Reflection.Metadata`
built-in without other dependencies | requires `System.Collections.Immutable`
`paket-files\fsprojects\FSharp.TypeProviders.SDK` | NuGet, may need ILMerge on some platforms
unknown cross-platform compatibility | net45, PCL, UAP, netstandard1.1, ...

I'd recommend to take #3239 (this), if/when/after known/proven safe on supported platforms, 
and replace it by #3223 if applicable, under the same conditions, to move towards net/standard