@mrinaldi 2018-06-07
This branch was lying around for quite some time waiting for tests to be promoted to a PR.
TBH, I didn't see the #3223 before or went through the PR list at all.
I did try #3223 , though and the performance figures as very similar to the ones on this PR.
In fact, the numbers are so similar that I had to double check if I was running the correct `paket.exe`.

Anyway, nice work @viktor-svub