@matthid 2018-06-10
I'm not sure this was the correct approach: It seems this `AssemblyReader.fs` thing doesn't actually exist anymore: https://github.com/fsprojects/FSharp.TypeProviders.SDK/tree/master/src
We would need to reference the complete type provider sdk in order to stay up-to-date. Therefore I think we should switch to @viktor-svub approach and use the microsoft package (or update to full type provider sdk).

In any case we should indeed go the complete way and remove mono.cecil instead of having multiple dependencies.

@viktor-svub Is it possible for you to update #3223? Or what do you think?