@mrinaldi 2018-06-11
I'm not sure what exactly @matthid is asking @viktor-svub to update. It might be this, but I'll point out anyway in case it's not.

`AssemblyReader` is used in a couple or places throughout the code base. In fact, this was the reason I used it instead of something else.

That said, I think we should aim to remove it completely in case we go with the microsoft package for the points @matthid already pointed out. @viktor-svub let me know if you need any help with that.